% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Estimates of Station usage}
\format{
A data.table with the following columns:
\describe{
\item{NLC}{National Location Code}
\item{year}{Financial year}
\item{metric}{One in: \code{Full}, \code{Reduced}, \code{Season}, \code{Changes}, \code{Total}, \code{Rank}, \code{VarYoY}}
\item{value}{The value for the Metric}
}
}
\usage{
dataset
}
\description{
Annual estimates of the number of Entries/Exits (Totals plus Full/Reduced/Season tickets),
and Interchanges at each Station in Great Britain.
}
\keyword{datasets}
