% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distances}
\alias{distances}
\title{Distances}
\format{
A data.table with the following columns:
\describe{
\item{NLCa}{National Location Code}
\item{NLCb}{National Location Code}
\item{distance}{Linear distance in metres}
\item{knn}{The order of neighborhood}
}
}
\usage{
distances
}
\description{
Linear Distances between each pair of Stations
}
\keyword{datasets}
