% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RstationsusageUK.R
\docType{package}
\name{RstationsusageUK}
\alias{RstationsusageUK}
\title{RstationsusageUK}
\description{
Annual estimates of the number of Entries/Exits (Totals plus Full/Reduced/Season tickets),
and Interchanges at each Station in Great Britain.
}
